Fix warning in GFAWriter

+++ assembler/src/common/assembly_graph/graph_support/contig_output.hpp
@@ -156,7 +156,7 @@ private:
     void WriteSegments(std::ofstream &stream) {
         GFASegmentWriter segment_writer(stream);
         for (auto it = graph_.ConstEdgeBegin(true); !it.IsEnd(); ++it) {
-            segment_writer.Write((*it).int_id(), graph_.EdgeNucls(*it), graph_.coverage(*it) * graph_.length(*it));
+            segment_writer.Write((*it).int_id(), graph_.EdgeNucls(*it), graph_.coverage(*it) * double(graph_.length(*it)));
         }
     }
